+++ AzeriteUI/back-end/plugins/generic_reputation.lua
@@ -1,6 +1,9 @@
 local LibClientBuild = Wheel("LibClientBuild")
 assert(LibClientBuild, "GenericReputation requires LibClientBuild to be loaded.")
 
+local LibNumbers = Wheel("LibNumbers")
+assert(LibNumbers, "GenericReputation requires LibNumbers to be loaded.")
+
 -- Lua API
 local _G = _G
 local math_floor = math.floor
@@ -20,36 +23,9 @@ local IsFactionParagon = C_Reputation and C_Reputation.IsFactionParagon
 local IsClassic = LibClientBuild:IsClassic()
 local IsRetail = LibClientBuild:IsRetail()
 
--- Number abbreviations
-local short = function(value)
-	value = tonumber(value)
-	if (not value) then return "" end
-	if (value >= 1e9) then
-		return ("%.1fb"):format(value / 1e9):gsub("%.?0+([kmb])$", "%1")
-	elseif value >= 1e6 then
-		return ("%.1fm"):format(value / 1e6):gsub("%.?0+([kmb])$", "%1")
-	elseif value >= 1e3 or value <= -1e3 then
-		return ("%.1fk"):format(value / 1e3):gsub("%.?0+([kmb])$", "%1")
-	else
-		return tostring(math_floor(value))
-	end	
-end
-
--- zhCN exceptions
-local gameLocale = GetLocale()
-if (gameLocale == "zhCN") then 
-	short = function(value)
-		value = tonumber(value)
-		if (not value) then return "" end
-		if (value >= 1e8) then
-			return ("%.1f亿"):format(value / 1e8):gsub("%.?0+([km])$", "%1")
-		elseif value >= 1e4 or value <= -1e3 then
-			return ("%.1f万"):format(value / 1e4):gsub("%.?0+([km])$", "%1")
-		else
-			return tostring(math_floor(value))
-		end 
-	end
-end 
+-- Number Abbreviation
+local short = LibNumbers:GetNumberAbbreviationShort()
+local large = LibNumbers:GetNumberAbbreviationLong()
 
 local UpdateValue = function(element, current, min, max, factionName, standingID, standingLabel, isFriend)
 	local value = element.Value or element:IsObjectType("FontString") and element 
@@ -214,5 +190,5 @@ end
 
 -- Register it with compatible libraries
 for _,Lib in ipairs({ (Wheel("LibUnitFrame", true)), (Wheel("LibNamePlate", true)), (Wheel("LibMinimap", true)) }) do 
-	Lib:RegisterElement("Reputation", Enable, Disable, Proxy, 2)
+	Lib:RegisterElement("Reputation", Enable, Disable, Proxy, 3)
 end