+++ AzeriteUI/back-end/plugins/generic_xp.lua
@@ -1,4 +1,7 @@
 
+local LibNumbers = Wheel("LibNumbers")
+assert(LibNumbers, "GenericXP requires LibNumbers to be loaded.")
+
 -- Lua API
 local _G = _G
 local math_floor = math.floor
@@ -7,43 +10,16 @@ local tonumber = tonumber
 local tostring = tostring
 
 -- WoW API
-local GetRestState = _G.GetRestState
-local GetTimeToWellRested = _G.GetTimeToWellRested
-local GetXPExhaustion = _G.GetXPExhaustion
-local IsResting = _G.IsResting
-local UnitXP = _G.UnitXP
-local UnitXPMax = _G.UnitXPMax
+local GetRestState = GetRestState
+local GetTimeToWellRested = GetTimeToWellRested
+local GetXPExhaustion = GetXPExhaustion
+local IsResting = IsResting
+local UnitXP = UnitXP
+local UnitXPMax = UnitXPMax
 
--- Number abbreviations
-local short = function(value)
-	value = tonumber(value)
-	if (not value) then return "" end
-	if (value >= 1e9) then
-		return ("%.1fb"):format(value / 1e9):gsub("%.?0+([kmb])$", "%1")
-	elseif value >= 1e6 then
-		return ("%.1fm"):format(value / 1e6):gsub("%.?0+([kmb])$", "%1")
-	elseif value >= 1e3 or value <= -1e3 then
-		return ("%.1fk"):format(value / 1e3):gsub("%.?0+([kmb])$", "%1")
-	else
-		return tostring(math_floor(value))
-	end	
-end
-
--- zhCN exceptions
-local gameLocale = GetLocale()
-if (gameLocale == "zhCN") then 
-	short = function(value)
-		value = tonumber(value)
-		if (not value) then return "" end
-		if (value >= 1e8) then
-			return ("%.1f亿"):format(value / 1e8):gsub("%.?0+([km])$", "%1")
-		elseif value >= 1e4 or value <= -1e3 then
-			return ("%.1f万"):format(value / 1e4):gsub("%.?0+([km])$", "%1")
-		else
-			return tostring(math_floor(value))
-		end 
-	end
-end 
+-- Number Abbreviation
+local short = LibNumbers:GetNumberAbbreviationShort()
+local large = LibNumbers:GetNumberAbbreviationLong()
 
 local UpdateValue = function(element, min, max, restedLeft, restedTimeLeft)
 	local value = element.Value or element:IsObjectType("FontString") and element 
@@ -200,5 +176,5 @@ end
 
 -- Register it with compatible libraries
 for _,Lib in ipairs({ (Wheel("LibUnitFrame", true)), (Wheel("LibNamePlate", true)), (Wheel("LibMinimap", true)) }) do 
-	Lib:RegisterElement("XP", Enable, Disable, Proxy, 12)
+	Lib:RegisterElement("XP", Enable, Disable, Proxy, 13)
 end