+++ AzeriteUI/back-end/plugins/unit_altpower.lua
@@ -6,6 +6,9 @@ if (LibClientBuild:IsClassic()) then
 	return
 end
 
+local LibNumbers = Wheel("LibNumbers")
+assert(LibNumbers, "UnitAltPower requires LibNumbers to be loaded.")
+
 -- Lua API
 local _G = _G
 local math_floor = math.floor
@@ -21,6 +24,10 @@ local UnitAlternatePowerInfo = UnitAlternatePowerInfo
 local UnitPower = UnitPower
 local UnitPowerBarID = UnitPowerBarID
 local UnitPowerMax = UnitPowerMax
+
+-- Number Abbreviation
+local short = LibNumbers:GetNumberAbbreviationShort()
+local large = LibNumbers:GetNumberAbbreviationLong()
 
 -- Sourced from BlizzardInterfaceResources/Resources/EnumerationTables.lua
 local ALTERNATE_POWER_INDEX = Enum and Enum.PowerType.Alternate or ALTERNATE_POWER_INDEX or 10
@@ -43,48 +50,6 @@ if (not UnitAlternatePowerInfo) then
 		return GetAlternatePowerInfoByID(barID)
 	end
 end
-
--- Number abbreviations
----------------------------------------------------------------------	
-local large = function(value)
-	if (value >= 1e8) then 		return string_format("%.0fm", value/1e6) 	-- 100m, 1000m, 2300m, etc
-	elseif (value >= 1e6) then 	return string_format("%.1fm", value/1e6) 	-- 1.0m - 99.9m 
-	elseif (value >= 1e5) then 	return string_format("%.0fk", value/1e3) 	-- 100k - 999k
-	elseif (value >= 1e3) then 	return string_format("%.1fk", value/1e3) 	-- 1.0k - 99.9k
-	elseif (value > 0) then 	return value 								-- 1 - 999
-	else 						return ""
-	end 
-end 
-
-local short = function(value)
-	value = tonumber(value)
-	if (not value) then return "" end
-	if (value >= 1e9) then
-		return ("%.1fb"):format(value / 1e9):gsub("%.?0+([kmb])$", "%1")
-	elseif value >= 1e6 then
-		return ("%.1fm"):format(value / 1e6):gsub("%.?0+([kmb])$", "%1")
-	elseif value >= 1e3 or value <= -1e3 then
-		return ("%.1fk"):format(value / 1e3):gsub("%.?0+([kmb])$", "%1")
-	else
-		return tostring(math_floor(value))
-	end	
-end
-
--- zhCN exceptions
-local gameLocale = GetLocale()
-if (gameLocale == "zhCN") then 
-	short = function(value)
-		value = tonumber(value)
-		if (not value) then return "" end
-		if (value >= 1e8) then
-			return ("%.1f亿"):format(value / 1e8):gsub("%.?0+([km])$", "%1")
-		elseif value >= 1e4 or value <= -1e3 then
-			return ("%.1f万"):format(value / 1e4):gsub("%.?0+([km])$", "%1")
-		else
-			return tostring(math_floor(value))
-		end 
-	end
-end 
 
 -- Borrow the unitframe tooltip
 local GetTooltip = function(element)
@@ -228,5 +193,5 @@ end
 
 -- Register it with compatible libraries
 for _,Lib in ipairs({ (Wheel("LibUnitFrame", true)), (Wheel("LibNamePlate", true)) }) do
-	Lib:RegisterElement("AltPower", Enable, Disable, Proxy, 15)
+	Lib:RegisterElement("AltPower", Enable, Disable, Proxy, 16)
 end