+++ AzeriteUI/front-end/modules/minimap.lua
@@ -4,6 +4,9 @@ if (not Core) then
 	return 
 end
 
+local LibNumbers = Wheel("LibNumbers")
+assert(LibNumbers, ADDON..":Minimap requires LibNumbers to be loaded.")
+
 local L = Wheel("LibLocale"):GetLocale(ADDON)
 local Module = Core:NewModule("Minimap", "LibEvent", "LibDB", "LibFrame", "LibMinimap", "LibTooltip", "LibTime", "LibSound", "LibPlayerData", "LibClientBuild")
 
@@ -42,6 +45,10 @@ local UnitExists = UnitExists
 local UnitLevel = UnitLevel
 local UnitRace = UnitRace
 
+-- Number Abbreviation
+local short = LibNumbers:GetNumberAbbreviationShort()
+local large = LibNumbers:GetNumberAbbreviationLong()
+
 -- Private API
 local Colors = Private.Colors
 local GetConfig = Private.GetConfig
@@ -77,36 +84,6 @@ local getTimeStrings = function(h, m, suffix, useStandardTime, abbreviateSuffix)
 	else 
 		return "%02.0f:%02.0f", h, m
 	end 
-end 
-
-local short = function(value)
-	value = tonumber(value)
-	if (not value) then return "" end
-	if (value >= 1e9) then
-		return ("%.1fb"):format(value / 1e9):gsub("%.?0+([kmb])$", "%1")
-	elseif (value >= 1e6) then
-		return ("%.1fm"):format(value / 1e6):gsub("%.?0+([kmb])$", "%1")
-	elseif (value >= 1e3) or (value <= -1e3) then
-		return ("%.1fk"):format(value / 1e3):gsub("%.?0+([kmb])$", "%1")
-	else
-		return tostring(math_floor(value))
-	end	
-end
-
--- zhCN exceptions
-local gameLocale = GetLocale()
-if (gameLocale == "zhCN") then 
-	short = function(value)
-		value = tonumber(value)
-		if (not value) then return "" end
-		if (value >= 1e8) then
-			return ("%.1f亿"):format(value / 1e8):gsub("%.?0+([km])$", "%1")
-		elseif (value >= 1e4) or (value <= -1e3) then
-			return ("%.1f万"):format(value / 1e4):gsub("%.?0+([km])$", "%1")
-		else
-			return tostring(math_floor(value))
-		end 
-	end
 end 
 
 local MouseIsOver = function(frame)