+++ AzeriteUI/front-end/private.lua
@@ -47,6 +47,9 @@ assert(LibColorTool, ADDON.." requires LibColorTool to be loaded.")
 local LibFontTool = Wheel("LibFontTool")
 assert(LibFontTool, ADDON.." requires LibFontTool to be loaded.")
 
+local LibNumbers = Wheel("LibNumbers")
+assert(LibNumbers, ADDON.." requires LibNumbers to be loaded.")
+
 -- Lua API
 local _G = _G
 local math_ceil = math.ceil
@@ -78,6 +81,10 @@ local UnitIsPlayer = UnitIsPlayer
 local UnitIsUnit = UnitIsUnit
 local UnitLevel = UnitLevel
 local UnitPowerMax = UnitPowerMax
+
+-- Number Abbreviation
+local short = LibNumbers:GetNumberAbbreviationShort()
+local large = LibNumbers:GetNumberAbbreviationLong()
 
 -- Addon localization
 local L = Wheel("LibLocale"):GetLocale(ADDON)
@@ -115,43 +122,12 @@ local degreesToRadians = function(degrees)
 	return degrees*deg2rad
 end 
 
-
 local getTimeStrings = function(h, m, suffix, useStandardTime, abbreviateSuffix)
 	if (useStandardTime) then 
 		return "%.0f:%02.0f |cff888888%s|r", h, m, abbreviateSuffix and string_match(suffix, "^.") or suffix
 	else 
 		return "%02.0f:%02.0f", h, m
 	end 
-end 
-
-local short = function(value)
-	value = tonumber(value)
-	if (not value) then return "" end
-	if (value >= 1e9) then
-		return ("%.1fb"):format(value / 1e9):gsub("%.?0+([kmb])$", "%1")
-	elseif (value >= 1e6) then
-		return ("%.1fm"):format(value / 1e6):gsub("%.?0+([kmb])$", "%1")
-	elseif (value >= 1e3) or (value <= -1e3) then
-		return ("%.1fk"):format(value / 1e3):gsub("%.?0+([kmb])$", "%1")
-	else
-		return tostring(math_floor(value))
-	end	
-end
-
--- zhCN exceptions
-local gameLocale = GetLocale()
-if (gameLocale == "zhCN") then 
-	short = function(value)
-		value = tonumber(value)
-		if (not value) then return "" end
-		if (value >= 1e8) then
-			return ("%.1f亿"):format(value / 1e8):gsub("%.?0+([km])$", "%1")
-		elseif (value >= 1e4) or (value <= -1e3) then
-			return ("%.1f万"):format(value / 1e4):gsub("%.?0+([km])$", "%1")
-		else
-			return tostring(math_floor(value))
-		end 
-	end
 end 
 
 -----------------------------------------------------------------